Fix up tests from today's pair programming session.
