+++ test/unit/subcommands/test_subcommands_listcmd.py
@@ -29,36 +29,23 @@
 
 import unittest
 import pytest
+from mock import patch, Mock
+from ..util import capture
+from hod.subcommands.listcmd import ListSubCommand
 
-from hod.subcommands.pbs import CreatePbsApplication
+class TestListSubCommand(unittest.TestCase):
+    def test_run(self):
+        app = ListSubCommand()
+        app.run([])
 
-from mock import patch
-
-class TestCreatePbsApplication(unittest.TestCase):
-    @pytest.mark.xfail(reason="Requires easybuild environment")
-    def test_run_no_args(self):
-        with patch('hod.subcommands.pbs.PbsHodJob'):
-            app = CreatePbsApplication()
-            self.assertRaises(ValueError, app.run, [])
-
-    @pytest.mark.xfail(reason="Requires easybuild environment")
-    def test_run_with_args(self):
-        with patch('hod.subcommands.pbs.PbsHodJob'):
-            app = CreatePbsApplication()
-            app.run(['--config=hod.conf', '--workdir=workdir'])
-
-    def test_run_with_dist_arg(self):
-        with patch('hod.subcommands.pbs.PbsHodJob'):
-            app = CreatePbsApplication()
-            app.run(['--dist=Hadoop-2.3.0', '--workdir=workdir'])
-
-    def test_run_fails_with_config_and_dist_arg(self):
-        with patch('hod.subcommands.pbs.PbsHodJob'):
-            app = CreatePbsApplication()
-            self.assertRaises(ValueError, app.run, 
-                    ['--config=hod.conf', '--dist=Hadoop-2.3.0', '--workdir=workdir'])
+    def test_run_good(self):
+        with patch('hod.rmscheduler.rm_pbs.Pbs', return_value=Mock(state=lambda: 'good')):
+            app = ListSubCommand()
+            with capture(app.run, []) as output:
+                self.assertEqual(output, 'good\n')
 
     def test_usage(self):
-        app = CreatePbsApplication()
+        app = ListSubCommand()
         usage = app.usage()
         self.assertTrue(isinstance(usage, basestring))
+
